+++ packages/react-universal/src/hooks/useNavigating.ts
@@ -1,15 +1,13 @@
 import React from "react";
 import useRouter from "./useRouter";
 
-import { CancelActiveNavigation } from "@curi/types";
-
 export default function useNavigating() {
   const router = useRouter();
   const [cancel, setCancel] = React.useState(undefined);
 
   React.useEffect(() => {
     let removed = false;
-    const stop = router.cancel((callback: CancelActiveNavigation) => {
+    const stop = router.cancel(callback => {
       if (!removed) {
         setCancel(() => callback);
       }