+++ packages/router/src/index.ts
@@ -1,6 +1,6 @@
 import { PrepareRoutesOptions } from "./prepare/prepareRoutes";
-
-export { PrepareRoutesOptions };
+import { RouterOptions } from "./router/createRouter";
+export { PrepareRoutesOptions, RouterOptions };
 
 import createRouter from "./router/createRouter";
 import prepareRoutes from "./prepare/prepareRoutes";