+++ packages/router/types/router/createRouter.d.ts
@@ -1,3 +1,9 @@
 import { HistoryConstructor, HistoryOptions } from "@hickory/root";
-import { RouteMatcher, CuriRouter, RouterOptions } from "@curi/types";
+import { RouteMatcher, CuriRouter, Observer } from "@curi/types";
+export interface RouterOptions<O = HistoryOptions> {
+    sideEffects?: Array<Observer>;
+    invisibleRedirects?: boolean;
+    external?: any;
+    history?: O;
+}
 export default function createRouter<O = HistoryOptions>(historyConstructor: HistoryConstructor<O>, routes: RouteMatcher, options?: RouterOptions<O>): CuriRouter;